Add files via upload
This is the first part of the code. I have to implement the new reading part of the code as per the new input file code as mentioned by you during our previous meeting.
As and when I get time between exam preparations, I shall keep updating the above code.